+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Le breadcrumb pouvait faire planter le navigateur s'il y avait trop d'entités #1321
 - Correction d'un appel déprécié dans FluxDataSedaHelios #1308
+- Fast TdT : Permettre la récupération d'un acquittement PES lorsque le dernier état est "Classé" ou "Archivé" #1325
+- Fast Parapheur : Permettre la récupération de la signature lorsque le dernier état est "Signé" ou "Archivé" #1325
 
 ## Évolutions